Partially restructured project to centralize library lib and include folders.

+++ 7DaysWorldEngine/Scene.h
@@ -18,5 +18,5 @@ public:
 	Renderer renderer;
 	StaticShader shader;
 	Entity entity;
-	int c = 1;
+	float c = 1;
 };
